requirements.txt
numpy>=1.21.0
tensorflow>=2.8.0
torch>=1.10.0
torchvision>=0.11.0
matplotlib>=3.5.0
seaborn>=0.11.2
pandas>=1.3.0
scikit-learn>=1.0.0
streamlit>=1.6.0
plotly>=5.3.1
pillow>=8.4.0
tqdm>=4.62.0
scipy>=1.7.0
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.1
